Fix sql

+++ app/src/main/resources/schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE urls (
 
 CREATE TABLE url_checks (
     id SERIAL PRIMARY KEY,
-    url_id bigint REFERENCES urls (id),
+    url_id INTEGER REFERENCES urls (id),
     status_code INTEGER NOT NULL,
     h1 VARCHAR(255),
     title VARCHAR(255),
